requirements.txt
flask
torch
torchvision
pillow
efficientnet_pytorch
timm
opencv-python
matplotlib
plotly
google-generativeai
numpy